If there's no data file at startup, create an empty file. Properly save file table changes  with MQTT updates (in both in memory and on disk modes). Add a helper script that can send an mqtt message (useful for adding data to a locally running server).
